requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
alembic==1.14.1
    # via flask-migrate
amqp==5.3.1
    # via kombu
aniso8601==10.0.0
    # via flask-restx
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
blinker==1.9.0
    # via
    #   flask
    #   flask-mail
    #   flask-principal
celery==5.4.0
    # via
    #   -r requirements.in
    #   flower
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via flask-security-too
faker==35.0.0
    # via -r requirements.in
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-login
    #   flask-mail
    #   flask-marshmallow
    #   flask-migrate
    #   flask-principal
    #   flask-restx
    #   flask-security-too
    #   flask-sqlalchemy
    #   flask-wtf
flask-login==0.6.3
    # via flask-security-too
flask-mail==0.10.0
    # via -r requirements.in
flask-marshmallow==1.3.0
    # via -r requirements.in
flask-migrate==4.1.0
    # via -r requirements.in
flask-principal==0.4.0
    # via flask-security-too
flask-restx==1.3.0
    # via -r requirements.in
flask-security-too==5.5.2
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.2.2
    # via flask-security-too
flower==2.0.1
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
humanize==4.11.0
    # via flower
idna==3.10
    # via email-validator
importlib-resources==6.5.2
    # via
    #   flask-restx
    #   flask-security-too
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.5
    # via flask
jsonschema==4.23.0
    # via flask-restx
jsonschema-specifications==2024.10.1
    # via jsonschema
kombu==5.4.2
    # via celery
lxml==5.3.0
    # via python-docx
mako==1.3.8
    # via alembic
markupsafe==3.0.2
    # via
    #   flask-security-too
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
marshmallow==3.26.0
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==1.4.0
    # via -r requirements.in
packaging==24.2
    # via marshmallow
passlib==1.7.4
    # via flask-security-too
peewee==3.17.8
    # via -r requirements.in
prometheus-client==0.21.1
    # via flower
prompt-toolkit==3.0.50
    # via click-repl
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   faker
python-docx==1.1.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
pytz==2024.2
    # via
    #   flask-restx
    #   flower
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.37
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
tornado==6.4.2
    # via flower
typing-extensions==4.12.2
    # via
    #   alembic
    #   faker
    #   python-docx
    #   sqlalchemy
tzdata==2025.1
    # via
    #   celery
    #   kombu
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
    #   flask-restx
wtforms==3.2.1
    # via
    #   flask-security-too
    #   flask-wtf
xlsxwriter==3.2.1
    # via -r requirements.in
